requirements.txt
discord.py==0.16.12
youtube_dl==2019.1.30.1
pynacl==1.3.0
